Fix UI display bugs
Multiple HostageRescue instances are simultaneously trying to seize the same global UI (RescueUIManager.Instance), and they keep hiding the UI from each other.

Add a static variable in HostageRescue:
public static HostageRescue current;
When a player enters the range of a hostage, set current to this hostage.
Only "current" is allowed to show/hide the UI.
When leaving, if the current one is being left, clear and hide the UI.
This way, it won't be hidden by the "Update" of other hostages.